write newly created vehicle to firebase
